update posts: post detail, adding new post
Rendering selected post is now available. With container
PostDetailContainer, id is passed in which is passed down from route's
params. Then calls for PostDetail comonent by giving an id.

Inside the PostDetail, it finds the exact post what user is looking for.
With custom fetchPost method, it gets post from posts of the
redux-store.

Also new post can be created with AddPost. Custom action is dispatched
when the form is submitted. The action calls corresponding login in
reducer and the reducer adds the new post object in the store.posts.